moved code from api.js to routes.js
